refactor version checker to internal location minor code improvements
