+++ cmd/kyma/kyma.go
@@ -40,7 +40,7 @@ Kyma CLI allows you to install and manage Kyma.
 		SilenceUsage:  true,
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
 			if !o.CI {
-				version.CheckForStableRelease()
+				cli.CheckForStableRelease(version.Version)
 			}
 		},
 		Run: func(cmd *cobra.Command, args []string) {
