+++ internal/cli/version.go
@@ -1,4 +1,4 @@
-package version
+package cli
 
 import (
 	"encoding/json"
@@ -10,15 +10,15 @@ import (
 )
 
 const (
-	GitHubAPIEndpoint = "https://api.github.com/repos/kyma-project/cli/releases/latest"
+	gitHubAPIEndpoint = "https://api.github.com/repos/kyma-project/cli/releases/latest"
 )
 
-type LatestGitTag struct {
+type latestGitTag struct {
 	Name string `json:"tag_name"`
 }
 
-func CheckForStableRelease() {
-	response, err := http.Get(GitHubAPIEndpoint)
+func CheckForStableRelease(currentVersion string) {
+	response, err := http.Get(gitHubAPIEndpoint)
 	if err != nil {
 		return
 	}
@@ -29,18 +29,19 @@ func CheckForStableRelease() {
 		return
 	}
 
-	var latestGitTag LatestGitTag
+	var latestGitTag latestGitTag
 	if err := json.Unmarshal(responseData, &latestGitTag); err != nil {
 		return
 	}
 
-	matched, err := regexp.MatchString("[0-9]+[.][0-9]+[.][0-9]+", Version)
-	if err != nil {
+	matched, err := regexp.MatchString("[0-9]+[.][0-9]+[.][0-9]+", currentVersion)
+	if err != nil || !matched {
 		return
-	} else if matched && Version < latestGitTag.Name {
+	}
+	if currentVersion < latestGitTag.Name {
 		nicePrint := nice.Nice{}
 		nicePrint.PrintImportantf("CAUTION: You're using an outdated version of the Kyma CLI (%s)."+
-			" The latest stable version is: %s", Version, latestGitTag.Name)
+			" The latest stable version is: %s", currentVersion, latestGitTag.Name)
 		fmt.Println()
 	}
 }
